Initial commit of comparison.js which merges 2 csvs.
Comparison.js is intended to be used to generate a single CSV file
which is produced by merging the output of two runs of index.js. This
new file can then be used as input to gnuplot for showing a direct
comparison between two sets of results.

Create csv.js with CSV handling code that's common to both
comparison.js and index.js.

Bug fixes in index.js.