Significant reworking.
Removed the idea of the 3d plot. It was extremely time-consuming to
produce and rather disappointing.

The previous code was difficult to extend to examine characteristics
other than extern- and linkonce-symbol performance. I'd really like to
be able to show performance for other metrics.

Another limitation was the assumption of two linkers being
compared. This version now allows an arbitrary number.